requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
load-dotenv==0.1.0
numpy==1.26.4
pandas==2.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
